+++ src/bn_generation.py
@@ -30,6 +30,7 @@ from __future__ import annotations
 import argparse
 from dataclasses import dataclass
 from typing import Any, Dict, List, Optional, Sequence, Tuple
+from pathlib import Path
 
 import numpy as np
 import networkx as nx
@@ -43,9 +44,13 @@ import matplotlib.pyplot as plt
 try:  # direct import when src/ is on sys.path
     from graph_generation import generate_dag_with_treewidth  # type: ignore
     from llm_calling import create_probability_prompt, run_llm_call, extract_numeric_answer  # type: ignore
+    from yaml_utils import load_yaml  # type: ignore
+    from cpd_utils import cpd_to_ascii_table  # type: ignore
 except ModuleNotFoundError:  # relative import when running as a module
     from .graph_generation import generate_dag_with_treewidth  # type: ignore
     from .llm_calling import create_probability_prompt, run_llm_call, extract_numeric_answer  # type: ignore
+    from .yaml_utils import load_yaml  # type: ignore
+    from .cpd_utils import cpd_to_ascii_table  # type: ignore
 
 
 # ------------------------------
@@ -306,103 +311,3 @@ if __name__ == "__main__":
     main()
 
 
-def _parse_field(val: Any) -> Any:
-    import ast, re
-    if isinstance(val, (list, dict)):
-        return val
-    if not isinstance(val, str):
-        return val
-    s = val.strip()
-    if not s:
-        return None
-    s = re.sub(r"np\.str_\(\'([^']*)\'\)", r"'\1'", s)
-    s = re.sub(r'np\.str_\("([^"]*)"\)', r"'\1'", s)
-    try:
-        return ast.literal_eval(s)
-    except Exception:
-        return val
-
-
-def inspect_row_and_call_llm(
-    *,
-    full_df: pd.DataFrame,
-    all_bayesian_networks: List[Dict[str, Any]],
-    row_index: int,
-    openai_client: Any,
-    model: str,
-    system_prompt: str = "You are a probability calculator. Provide exact numerical answers.",
-    draw_kwargs: Optional[Dict[str, Any]] = None,
-) -> Dict[str, Any]:
-    if draw_kwargs is None:
-        draw_kwargs = {}
-
-    row = full_df.iloc[row_index]
-    if "bn_index" not in row:
-        raise ValueError("full_df is missing 'bn_index' column")
-    bn_idx = int(row["bn_index"])  # type: ignore[arg-type]
-    bn = all_bayesian_networks[bn_idx]["bn"]
-
-    q_vars = _parse_field(row.get("query_vars"))
-    q_states = _parse_field(row.get("query_states"))
-    evidence = _parse_field(row.get("evidence"))
-    if not isinstance(q_vars, list) or not isinstance(q_states, list):
-        raise ValueError("Row has invalid 'query_vars' or 'query_states'")
-    if evidence is not None and not isinstance(evidence, dict):
-        raise ValueError("Row has invalid 'evidence' (expected dict or None)")
-
-    # Draw BN
-    G = nx.DiGraph()
-    G.add_nodes_from(bn.nodes())
-    G.add_edges_from(bn.edges())
-    pos = nx.spring_layout(G, seed=0)
-    plt.figure(figsize=draw_kwargs.pop("figsize", (6, 4)))
-    nx.draw(G, pos, with_labels=True, node_color="#A7C7E7", arrows=True, **draw_kwargs)
-    plt.title("Bayesian Network Structure")
-    plt.show()
-
-    # Build query text
-    if len(q_vars) == 1:
-        if evidence:
-            ev_str = ", ".join([f"{k}={v}" for k, v in evidence.items()])
-            q_text = f"P({q_vars[0]}={q_states[0]} | {ev_str})"
-        else:
-            q_text = f"P({q_vars[0]}={q_states[0]})"
-    else:
-        parts = [f"{v}={s}" for v, s in zip(q_vars, q_states)]
-        if evidence:
-            ev_str = ", ".join([f"{k}={v}" for k, v in evidence.items()])
-            q_text = f"P({', '.join(parts)} | {ev_str})"
-        else:
-            q_text = f"P({', '.join(parts)})"
-    print("Query:", q_text)
-
-    exact_prob = row.get("probability", None)
-    if exact_prob is not None:
-        exact_prob = float(exact_prob)
-    print("Exact probability:", exact_prob)
-
-    prompt_str = create_probability_prompt(bn, q_vars, q_states, evidence)
-    messages = [
-        {"role": "system", "content": system_prompt},
-        {"role": "user", "content": prompt_str},
-    ]
-    try:
-        response, _ = run_llm_call(openai_client=openai_client, model=model, messages=messages)
-    except Exception as e:
-        print(f"Error calling LLM: {e}")
-        response = None
-    llm_prob = extract_numeric_answer(response) if response else None
-    print("LLM probability:", llm_prob)
-
-    delta = None
-    if llm_prob is not None and exact_prob is not None:
-        delta = float(abs(llm_prob - exact_prob))
-        print("Absolute error:", delta)
-
-    return {
-        "query": q_text,
-        "exact_probability": exact_prob,
-        "llm_probability": llm_prob,
-        "delta": delta,
-        "llm_response": response,
-    }
