+++ src/bn_query_sweep.py
@@ -1,8 +1,11 @@
 from __future__ import annotations
 
 from typing import Any, Dict, List, Optional, Sequence, Tuple
+from pathlib import Path
 
 import pandas as pd
+import matplotlib.pyplot as plt
+import networkx as nx
 
 from pgmpy.inference import VariableElimination
 
@@ -17,6 +20,7 @@ try:
         create_probability_prompt,
         extract_numeric_answer,
     )  # type: ignore
+    from yaml_utils import load_yaml  # type: ignore
 except ModuleNotFoundError:  # pragma: no cover
     from .graph_generation import generate_dag_with_treewidth  # type: ignore
     from .bn_generation import generate_variants_for_dag  # type: ignore
@@ -27,6 +31,7 @@ except ModuleNotFoundError:  # pragma: no cover
         create_probability_prompt,
         extract_numeric_answer,
     )
+    from .yaml_utils import load_yaml  # type: ignore
 
 
 def _arity_to_str(spec: Dict[str, Any]) -> str:
@@ -197,7 +202,8 @@ def call_llms_for_rows(
     all_bn_queries: List[List[QuerySpec]],
     openai_client: Any,
     model: str,
-    system_prompt: str = "You are a probability calculator. Provide exact numerical answers.",
+    prompts_path: Optional[Path] = None,
+    system_prompt: Optional[str] = None,
 ) -> pd.DataFrame:
     """Run LLM calls only for selected rows, augmenting full_df with LLM outputs.
 
@@ -206,6 +212,15 @@ def call_llms_for_rows(
     """
     import ast
     import re
+
+    # Load prompts from YAML if path provided
+    if prompts_path is not None:
+        prompts = load_yaml(prompts_path)
+        system_prompt = prompts.get("system_prompt", "You are a probability calculator. Provide exact numerical answers.")
+        prompt_template = prompts.get("prompt_base", "")
+    else:
+        system_prompt = system_prompt or "You are a probability calculator. Provide exact numerical answers."
+        prompt_template = ""
 
     def _maybe_parse(val):
         # If already a structure, return as-is
@@ -252,7 +267,34 @@ def call_llms_for_rows(
             dict(evidence_val) if isinstance(evidence_val, dict) else None
         )
 
-        prompt_str = create_probability_prompt(bn, query_vars, query_states, evidence)
+        # Create prompt using YAML template if available, otherwise fallback to create_probability_prompt
+        if prompt_template:
+            # Generate CPTs as ASCII tables
+            cpd_strings = []
+            for cpd in bn.get_cpds():
+                cpd_strings.append(cpd_to_ascii_table(cpd))
+            cpds_as_string = "\n\n".join(cpd_strings)
+            
+            # Create query string
+            if len(query_vars) == 1:
+                if evidence:
+                    ev_str = ", ".join([f"{k}={v}" for k, v in evidence.items()])
+                    query_str = f"P({query_vars[0]}={query_states[0]} | {ev_str})"
+                else:
+                    query_str = f"P({query_vars[0]}={query_states[0]})"
+            else:
+                parts = [f"{v}={s}" for v, s in zip(query_vars, query_states)]
+                if evidence:
+                    ev_str = ", ".join([f"{k}={v}" for k, v in evidence.items()])
+                    query_str = f"P({', '.join(parts)} | {ev_str})"
+                else:
+                    query_str = f"P({', '.join(parts)})"
+            
+            # Format the prompt using the template
+            prompt_str = prompt_template.format(cpts=cpds_as_string, query=query_str)
+        else:
+            prompt_str = create_probability_prompt(bn, query_vars, query_states, evidence)
+        
         messages = [
             {"role": "system", "content": system_prompt},
             {"role": "user", "content": prompt_str},
@@ -272,8 +314,169 @@ def call_llms_for_rows(
     return out
 
 
+def _parse_field(val: Any) -> Any:
+    """Parse field values that might be stringified lists/dicts from DataFrame storage."""
+    import ast
+    import re
+    
+    if isinstance(val, (list, dict)):
+        return val
+    if not isinstance(val, str):
+        return val
+    s = val.strip()
+    if not s:
+        return None
+    # Normalize occurrences like np.str_('X') -> 'X'
+    s = re.sub(r"np\.str_\(\'([^']*)\'\)", r"'\1'", s)
+    # Also the double-quoted variant
+    s = re.sub(r'np\.str_\("([^"]*)"\)', r"'\1'", s)
+    try:
+        return ast.literal_eval(s)
+    except Exception:
+        # Fallback: return original string
+        return val
+
+
+def inspect_row_and_call_llm(
+    *,
+    full_df: pd.DataFrame,
+    all_bayesian_networks: List[Dict[str, Any]],
+    row_index: int,
+    openai_client: Any,
+    model: str,
+    prompts_path: Optional[Path] = None,
+    system_prompt: Optional[str] = None,
+    draw_kwargs: Optional[Dict[str, Any]] = None,
+    print_prompts: bool = True,
+) -> Dict[str, Any]:
+    """Inspect a specific row from full_df: retrieve BN and query, draw BN, print query, call LLM, and compare results.
+    
+    Args:
+        full_df: DataFrame containing BN and query information
+        all_bayesian_networks: List of BN dictionaries with 'bn' and 'meta' keys
+        row_index: Index of the row in full_df to inspect
+        openai_client: Initialized OpenAI client
+        model: Model name to use
+        prompts_path: Path to prompts.yaml file (optional)
+        system_prompt: Custom system prompt (optional, overrides prompts_path)
+        draw_kwargs: Additional arguments for BN drawing (optional)
+        print_prompts: Whether to print system and user prompts to console (default: True)
+    
+    Returns:
+        Dictionary with query, exact probability, LLM probability, error, and LLM response
+    """
+    if draw_kwargs is None:
+        draw_kwargs = {}
+
+    row = full_df.iloc[row_index]
+    if "bn_index" not in row:
+        raise ValueError("full_df is missing 'bn_index' column")
+    bn_idx = int(row["bn_index"])  # type: ignore[arg-type]
+    bn = all_bayesian_networks[bn_idx]["bn"]
+
+    q_vars = _parse_field(row.get("query_vars"))
+    q_states = _parse_field(row.get("query_states"))
+    evidence = _parse_field(row.get("evidence"))
+    if not isinstance(q_vars, list) or not isinstance(q_states, list):
+        raise ValueError("Row has invalid 'query_vars' or 'query_states'")
+    if evidence is not None and not isinstance(evidence, dict):
+        raise ValueError("Row has invalid 'evidence' (expected dict or None)")
+
+    # Draw BN
+    G = nx.DiGraph()
+    G.add_nodes_from(bn.nodes())
+    G.add_edges_from(bn.edges())
+    pos = nx.spring_layout(G, seed=0)
+    plt.figure(figsize=draw_kwargs.pop("figsize", (6, 4)))
+    nx.draw(G, pos, with_labels=True, node_color="#A7C7E7", arrows=True, **draw_kwargs)
+    plt.title("Bayesian Network Structure")
+    plt.show()
+
+    # Build query text
+    if len(q_vars) == 1:
+        if evidence:
+            ev_str = ", ".join([f"{k}={v}" for k, v in evidence.items()])
+            q_text = f"P({q_vars[0]}={q_states[0]} | {ev_str})"
+        else:
+            q_text = f"P({q_vars[0]}={q_states[0]})"
+    else:
+        parts = [f"{v}={s}" for v, s in zip(q_vars, q_states)]
+        if evidence:
+            ev_str = ", ".join([f"{k}={v}" for k, v in evidence.items()])
+            q_text = f"P({', '.join(parts)} | {ev_str})"
+        else:
+            q_text = f"P({', '.join(parts)})"
+    print("Query:", q_text)
+
+    exact_prob = row.get("probability", None)
+    if exact_prob is not None:
+        exact_prob = float(exact_prob)
+    print("Exact probability:", exact_prob)
+
+    # Load prompts from YAML if path provided
+    if prompts_path is not None:
+        prompts = load_yaml(prompts_path)
+        system_prompt = prompts.get("system_prompt", "You are a probability calculator. Provide exact numerical answers.")
+        prompt_template = prompts.get("prompt_base", "")
+    else:
+        system_prompt = system_prompt or "You are a probability calculator. Provide exact numerical answers."
+        prompt_template = ""
+
+    # Create prompt using YAML template if available, otherwise fallback to create_probability_prompt
+    if prompt_template:
+        # Generate CPTs as ASCII tables
+        cpd_strings = []
+        for cpd in bn.get_cpds():
+            cpd_strings.append(cpd_to_ascii_table(cpd))
+        cpds_as_string = "\n\n".join(cpd_strings)
+        
+        # Format the prompt using the template
+        prompt_str = prompt_template.format(cpts=cpds_as_string, query=q_text)
+    else:
+        prompt_str = create_probability_prompt(bn, q_vars, q_states, evidence)
+
+    messages = [
+        {"role": "system", "content": system_prompt},
+        {"role": "user", "content": prompt_str},
+    ]
+
+    # Print prompts to console if requested
+    if print_prompts:
+        print("\n" + "="*50)
+        print("SYSTEM PROMPT:")
+        print("="*50)
+        print(system_prompt)
+        print("\n" + "="*50)
+        print("USER PROMPT:")
+        print("="*50)
+        print(prompt_str)
+        print("="*50 + "\n")
+
+    try:
+        response, _ = run_llm_call(openai_client=openai_client, model=model, messages=messages)
+    except Exception as e:
+        print(f"Error calling LLM: {e}")
+        response = None
+    llm_prob = extract_numeric_answer(response) if response else None
+    print("LLM probability:", llm_prob)
+
+    delta = None
+    if llm_prob is not None and exact_prob is not None:
+        delta = float(abs(llm_prob - exact_prob))
+        print("Absolute error:", delta)
+
+    return {
+        "query": q_text,
+        "exact_probability": exact_prob,
+        "llm_probability": llm_prob,
+        "delta": delta,
+        "llm_response": response,
+    }
+
+
 __all__ = [
     "generate_bn_sweep",
     "generate_queries_for_bn_list",
     "call_llms_for_rows",
+    "inspect_row_and_call_llm",
 ]
